Relocate `labeler` and docstring the `builder` module
